#2225 :truck: move obw component files in ergebnismeldung/OBW
